Change Bell and Interleaving into local functions [WIP]
